+++ installer/windows/Remove-FirewallRule.ps1
@@ -1,7 +1,0 @@
-[CmdletBinding()]
-param()
-
-$ErrorActionPreference = 'Stop'
-
-Get-NetFirewallRule -DisplayName 'LibreNMS Windows Agent TCP 6556' -ErrorAction SilentlyContinue |
-    Remove-NetFirewallRule